更新文档

+++ src/main/java/indi/yufr/jvm/yuShare/vm/oops/attribute/LineNumberTableAttribute.java
@@ -15,7 +15,7 @@ public class LineNumberTableAttribute extends AttributeInfo {
 
     @Data
     public static class Item {
-        private int startPc;
-        private int lineNumber;
+        private short startPc;
+        private short lineNumber;
     }
 }
